Add : Navbar, Post, ProfileSection 컴포넌트 추가 및 media query 추가
