+++ src/pages/jungwhanKim/Main/Components/Box/Box.scss
@@ -4,4 +4,18 @@
   margin-top: 40px;
   border-radius: 12px;
   background-color: #ffffff;
+
+  .box-header {
+    @include alignedSpaceBetween;
+    margin-bottom: 8px;
+
+    .header-text {
+      opacity: 0.6;
+      font-size: 16px;
+    }
+
+    .header-more {
+      font-size: 13px;
+    }
+  }
 }
