+++ src/pages/jungwhanKim/Main/Components/Comment/Comment.scss
@@ -20,24 +20,22 @@
   }
 
   .comment-like {
+    @include fullBgImage(
+      '../../../../../assets/jungwhanKim/heart-inactive.jpeg'
+    );
     width: 16px;
     height: 16px;
     margin-right: 16px;
     opacity: 0.5;
-    background-image: url('../../../../../assets/jungwhanKim/heart-inactive.jpeg');
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: cover;
   }
 
   .comment-like-clicked {
+    @include fullBgImage(
+      'https://www.pngitem.com/pimgs/m/522-5229528_transparent-background-instagram-heart-icon-hd-png-download.png'
+    );
     width: 16px;
     height: 16px;
     margin-right: 16px;
-    background-image: url('https://www.pngitem.com/pimgs/m/522-5229528_transparent-background-instagram-heart-icon-hd-png-download.png');
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: cover;
   }
 
   .delete-btn {
