+++ src/pages/jungwhanKim/Main/Components/ProfileSection/ProfileSection.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import './ProfileSection.scss';
+
+const ProfileSection = () => {
+  return (
+    <div className="profile-section">
+      <div className="profile-image" />
+      <div className="account-name">
+        <div className="account-name-main">_georgekim</div>
+        <div className="account-name-sub">44기 김정환</div>
+      </div>
+    </div>
+  );
+};
+
+export default ProfileSection;
